requirements.txt
Django==2.2.12
django-crispy-forms==1.9.1
django-material==0.5.1
django-shopping-cart==0.1
factory-boy==2.12.0
Faker==4.1.0
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.1.0
